+++ crates/zk-prover/src/circuits/dkg/mod.rs
@@ -4,4 +4,4 @@
 // without even the implied warranty of MERCHANTABILITY
 // or FITNESS FOR A PARTICULAR PURPOSE
 
-mod pkbfv;
+mod pk;
